refactor(audit): split upstream_workaround + comment_blocks into submodules
Follow-up to #1529 (which merged the original 540-line patch in
comment_hygiene.rs as-is). Address the audit findings the new code
brought into scope:

- comment_hygiene.rs hit high_item_count (27 top-level items, threshold 15)
- 3x intra_method_duplicate from repeated Finding { ... } construction
- missing_test_method for the public run() entry point

Move upstream-workaround code (Tier A marker+reference, Tier B version_compare)
to a new sibling module `upstream_workaround.rs`. Move the comment-block
extraction state machine to `comment_blocks.rs` so it can be reused by
other hygiene rules later. Both new files stay under the 15-item threshold.

Other improvements:
- Collapse WORKAROUND_MARKERS + WORKAROUND_PHRASES into a single
  MARKER_LITERALS catalogue.
- Collapse three URL regexes (GITHUB_ISSUE_PR, TRAC_TICKET, SEE_ISSUE_URL)
  into one alternation REFERENCE_RE.
- Extract make_finding(...) helper in comment_hygiene to remove the
  intra-method duplication around Finding construction.
- Add test_run() smoke test for the public entry point.
- Rename comment_blocks tests with extract_ prefix so the descriptive
  test-name matcher (#1518) recognizes them as covering extract().

Behaviour unchanged. 1471/0 lib tests serial. Self-audit clean
(drift_increased: False, 4 resolved findings, 0 new).
