+++ src/core/code_audit/mod.rs
@@ -14,6 +14,7 @@
 pub mod baseline;
 mod checks;
 pub mod codebase_map;
+mod comment_blocks;
 mod comment_hygiene;
 pub mod compare;
 mod compiler_warnings;
@@ -45,6 +46,7 @@ mod structural;
 mod test_coverage;
 pub(crate) mod test_mapping;
 mod test_topology;
+mod upstream_workaround;
 pub(crate) mod walker;
 mod wrapper_inference;
